+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add frontend service image management hooks and utilities with support for galleries
+- Integrate image processing utilities in service creation workflow with support for optional image uploads
 
 ### Added